[contracts] Move Shortest Path in a Grid to complete folder
