+++ src/contracts/Shortest-Path-in-a-Grid.ts
